requirements.txt
opencv-python>=4.8.0
numpy>=1.23.0
scipy>=1.10.0
Pillow>=10.0.0
scikit-image>=0.21.0
matplotlib>=3.6.0
pandas>=2.0.0
pyyaml>=6.0
scikit-learn>=1.4
# Optional for advanced quality metrics
# imquality[brisque]
# piq
